+++ core/execution/engine.go
@@ -364,7 +364,12 @@ func (e *Engine) succeedOrRestore(ctx context.Context, successor, parent string,
 	// restoring a market, but no state of the market nor parent market exists. Treat market as parent.
 	if restore && !sok && !ok {
 		// restoring a market, but the market state and parent market both are missing
+		// this market, upon leaving opening auction, cannot possibly succeed a market that no longer exists
+		// now we should reset
 		mkt.ResetParentIDAndInsurancePoolFraction()
+		// remove from maps
+		delete(e.successors, parent)
+		delete(e.isSuccessor, successor)
 		return nil
 	}
 	// if parent market is active, mark as succeeded
@@ -424,9 +429,11 @@ func (e *Engine) RestoreMarket(ctx context.Context, marketConfig *types.Market) 
 		if len(marketConfig.ParentMarketID) == 0 {
 			return nil
 		}
-		// check to see if the parent market can be found, remove if the parent market is gone
+		// check to see if the parent market can be found, remove from the successor maps if the parent is gone
+		// the market itself should still hold the reference because state was restored
 		if _, ok := e.futureMarkets[marketConfig.ParentMarketID]; !ok {
-			e.futureMarkets[marketConfig.ID].ResetParentIDAndInsurancePoolFraction()
+			delete(e.successors, marketConfig.ParentMarketID)
+			delete(e.isSuccessor, marketConfig.ID)
 		}
 		return nil
 	}